+++ src/DataType/DateTimeHandler.php
@@ -28,7 +28,7 @@ class DateTimeHandler implements HandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function canHandleValue($value): bool
+    public function canHandleValue(mixed $value): bool
     {
         return $value instanceof DateTimeInterface;
     }
@@ -36,7 +36,7 @@ class DateTimeHandler implements HandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function serializeValue($value): string
+    public function serializeValue(mixed $value): string
     {
         return $value->format($this->format);
     }
@@ -44,8 +44,8 @@ class DateTimeHandler implements HandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function unserializeValue(string $value)
+    public function unserializeValue(string $serializedValue): mixed
     {
-        return Carbon::createFromFormat($this->format, $value);
+        return Carbon::createFromFormat($this->format, $serializedValue);
     }
 }